fix(jql): Get missing issues
The previous jql query was missing:
* Subcomponents: Agent, ARO, OCPV, OpenStack, ROSA
* Was missing issues that were opened targeting older (and in the future
  newer) releases that might still be missing backports.

Of course, the resulting jql is not perfect either, as there can be bugs
that are marked as dependent to other bugs that they are not backports
of. That case is not too frequent, but I believe we could try to address
it in the future.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>